Merge...


Former-commit-id: f94fb9153108589fb5cb1771f68c9816b6c24f73

+++ invest_natcap/biodiversity/biodiversity_biophysical.py
@@ -98,7 +98,7 @@ def execute(args):
             'not match the columns in the sensitivity table')
 
     biophysical_args['half_saturation'] = \
-            int(args['half_saturation_constant'])    
+           float(args['half_saturation_constant'])    
 
     # if the access shapefile was provided add it to the dictionary
     try:
